second commit

+++ ImageLoading/imagesContorller.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace Localhost\ImageLoading;
 
 use Localhost\Service\DB;
@@ -27,8 +26,8 @@ class imagesContorller
         $allViewsData = $pdo->execute(self::$queryOne, $options); ?>
 
         <img src="../uploaded<?= DIRECTORY_SEPARATOR . $allViewsData['name'] ?>" class="pimg" title="<?= $file ?>"/>
-            <p align="center" >Просмотров: <?= $allViewsData['views'] ?></p>
-        <?php }
+        <p align="center">Просмотров: <?= $allViewsData['views'] ?></p>
+    <?php }
 
     public static function ShowMy()
     {
@@ -40,23 +39,25 @@ class imagesContorller
             foreach ($PicData as $rowPicData) { ?>
                 <table align="center" id="uploaded_image" border="2">
                     <thead bgcolor="#2F4F4F" style="color: #FFFFFF ">
-                        <td>
-                            <?php
-                            if (SessionManager::create()->user('user_id') === $rowPicData['user_id']) {
-                                echo SessionManager::create()->user('user_login'); } ?>
-                            <a href="deleteImage.php?id=<?=  $rowPicData['name'] ?>">Удалить<a>
-                        </td>
+                    <td>
+                        <?php
+                        if (SessionManager::create()->user('user_id') === $rowPicData['user_id']) {
+                            echo SessionManager::create()->user('user_login');
+                        } ?>
+                        <a href="deleteImage.php?id=<?= $rowPicData['name'] ?>">Удалить<a>
+                    </td>
                     </thead>
                     <tbody align="center" bgcolor="black">
-                        <tr>
-                            <td>
-                                <a href="details.php?id=<?= $rowPicData['id'] ?>">
-                                    <div id="uploaded_image" class="blok_img">
-                                        <img src="../uploaded/<?= $rowPicData['name']?>" width="500px" class="pimg" title="<?= $rowPicData['name'] ?>"/>
-                                    </div>
-                                </a>
-                            </td>
-                        </tr>
+                    <tr>
+                        <td>
+                            <a href="details.php?id=<?= $rowPicData['id'] ?>">
+                                <div id="uploaded_image" class="blok_img">
+                                    <img src="../uploaded/<?= $rowPicData['name'] ?>" width="500px" class="pimg"
+                                         title="<?= $rowPicData['name'] ?>"/>
+                                </div>
+                            </a>
+                        </td>
+                    </tr>
                     </tbody>
                 </table><br>
                 <?php
@@ -67,7 +68,7 @@ class imagesContorller
     public static function ShowUser()
     {
         $pdo = new DB();
-        $options = [':id' => $_GET['id'] ];
+        $options = [':id' => $_GET['id']];
         $PicData = $pdo->executeAll(self::$queryMy, $options);
 
         if (!empty(self::$dir)) {
@@ -77,7 +78,8 @@ class imagesContorller
                     <td>
                         <?php
                         if ($rowPicData['user_id'] === $rowPicData['user_login']) {
-                            echo SessionManager::create()->user('user_login'); } ?>
+                            echo SessionManager::create()->user('user_login');
+                        } ?>
                     </td>
                     </thead>
                     <tbody align="center" bgcolor="black">
@@ -85,7 +87,8 @@ class imagesContorller
                         <td>
                             <a href="details.php?id=<?= $rowPicData['name'] ?>">
                                 <div id="uploaded_image" class="blok_img">
-                                    <img src="<?= $rowPicData['name']  ?>" height="550px" class="pimg" title="<?= $rowPicData['name'] ?>"/>
+                                    <img src="<?= $rowPicData['name'] ?>" height="550px" class="pimg"
+                                         title="<?= $rowPicData['name'] ?>"/>
                                 </div>
                             </a>
                         </td>
@@ -110,15 +113,17 @@ class imagesContorller
                     <td>
                         <?php
                         if (SessionManager::create()->user('user_id') === $rowPicData['user_id']) {
-                            echo SessionManager::create()->user('user_login'); } ?>
+                            echo SessionManager::create()->user('user_login');
+                        } ?>
                     </td>
                     </thead>
                     <tbody align="center" bgcolor="black">
                     <tr>
                         <td>
-                            <a href="details.php?id=<?= $rowPicData['name'] ?>">
+                            <a href="ImageLoading/details.php?id=<?= $rowPicData['name'] ?>">
                                 <div id="uploaded_image" class="blok_img">
-                                    <img src="<?= $rowPicData['name']  ?>" height="550px" class="pimg" title="<?= $rowPicData['name'] ?>"/>
+                                    <img src="<?= $rowPicData['name'] ?>" height="550px" class="pimg"
+                                         title="<?= $rowPicData['name'] ?>"/>
                                 </div>
                             </a>
                         </td>
@@ -142,27 +147,31 @@ class imagesContorller
                 foreach ($allUsersData as $rowUsersData) { ?>
                     <table align="center" id="uploaded_image" border="2" width="650px" height="650px">
                     <thead bgcolor="#2F4F4F" style="color: #FFFFFF ">
-                        <?php if ($PicData['user_id'] === $rowUsersData['user_id']) { ?>
-                            <?php if ( $rowUsersData['user_login'] != SessionManager::create()->user('user_login') ){ ?>
-                        <td><a href="lk_form.php?id=<?= $rowUsersData['user_id']; } ?>"> <?= $rowUsersData['user_login'] ?></a></td>
-                        <?php } } ?>
-                    </thead>
-                        <tbody align="center" bgcolor="black">
-                            <tr>
-                                <td>
-                                    <a href="details.php?id=<?= $PicData['name'] ?>">
-                                        <div class="blok_img">
-                                            <img src="../uploaded/<?= $PicData['name'] ?>" height="550px" class="img" title="<?= $PicData['name']; ?>"/>
-                                        </div>
-                                    </a>
-                                    <?php foreach ($allViewsData as $rowViewData) {
-                                        if ($rowViewData['name'] == $PicData['name']) { ?>
-                                            <p id="views" style="color: white">Просмотров: <?= $rowViewData['views'] ?></p>
-                                        <?php } } ?>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table><br>
+                    <?php if ($PicData['user_id'] === $rowUsersData['user_id']) { ?>
+                        <?php if ($rowUsersData['user_login'] != SessionManager::create()->user('user_login')) { ?>
+                            <td><a href="lk_form.php?id=<?= $rowUsersData['user_id'];
+                        } ?>"> <?= $rowUsersData['user_login'] ?></a></td>
+                    <?php }
+                } ?>
+                </thead>
+                <tbody align="center" bgcolor="black">
+                <tr>
+                    <td>
+                        <a href="ImageLoading/details.php?id=<?= $PicData['id'] ?>">
+                            <div class="blok_img">
+                                <img src="../uploaded/<?= $PicData['name'] ?>" height="550px" class="img"
+                                     title="<?= $PicData['name']; ?>"/>
+                            </div>
+                        </a>
+                        <?php foreach ($allViewsData as $rowViewData) {
+                            if ($rowViewData['name'] == $PicData['name']) { ?>
+                                <p id="views" style="color: white">Просмотров: <?= $rowViewData['views'] ?></p>
+                            <?php }
+                        } ?>
+                    </td>
+                </tr>
+                </tbody>
+                </table><br>
             <?php }
         }
     }
@@ -171,7 +180,7 @@ class imagesContorller
     {
         $pdo = new DB();
         $id = $_GET['id'];
-        $options = [':id' => $_GET['id'] ];
+        $options = [':id' => $_GET['id']];
         $PicData = $pdo->executeAll(self::$queryMy, $options);
         $UsersData = $pdo->execute(self::$select_where, $options); ?>
         <table align="center">
@@ -186,7 +195,7 @@ class imagesContorller
                 <td>
                     <?php
                     echo $UsersData['user_login']; ?>
-                    <a href="deleteImage.php?id=<?=  $rowPicData['name'] ?>"><font color="red">Удалить</font><a>
+                    <a href="deleteImage.php?id=<?= $rowPicData['name'] ?>"><font color="red">Удалить</font><a>
                 </td>
                 </thead>
                 <tbody align="center" bgcolor="black">
@@ -194,7 +203,8 @@ class imagesContorller
                     <td>
                         <a href="details.php?id=<?= $rowPicData['name'] ?>">
                             <div id="uploaded_image" class="blok_img">
-                                <img src="<?= $rowPicData['name']  ?>" height="550px" class="pimg" title="<?= $rowPicData['name'] ?>"/>
+                                <img src="<?= $rowPicData['name'] ?>" height="550px" class="pimg"
+                                     title="<?= $rowPicData['name'] ?>"/>
                             </div>
                         </a>
                     </td>
@@ -219,25 +229,28 @@ class imagesContorller
                     <table align="center" id="uploaded_image" border="2" >
                     <thead bgcolor="#2F4F4F" style="color: #FFFFFF ">
                     <?php if ($PicData['user_id'] === $rowUsersData['user_id']) { ?>
-                        <?php if ( $PicData['user_id'] == SessionManager::create()->user('user_id') ){ ?>
-                        <td>
-                            <a href="lk_form.php">
-                                <?=SessionManager::create()->user('user_login')?>
-                            </a>
-                            <a href="deleteImage.php?id=<?=$PicData['name']?>">
-                                <font color="blue">Удалить</font>
-                            </a>
-                        </td>
-                            <?php } if ( $PicData['user_id'] !== SessionManager::create()->user('user_id') ){ ?>
-                        <td>
-                            <a href="lk_form.php?id=<?= $rowUsersData['user_id'] ?>">
-                                <font color="#f0f8ff"><?= $rowUsersData['user_login'], " " ?><font>
-                            </a>
-                            <a href="deleteImage.php?id=<?=  $PicData['name'] ?>">
-                                <font color="red">Удалить</font>
-                            </a>
-                        </td>
-                    <?php } } } ?>
+                        <?php if ($PicData['user_id'] == SessionManager::create()->user('user_id')) { ?>
+                            <td>
+                                <a href="lk_form.php">
+                                    <?= SessionManager::create()->user('user_login') ?>
+                                </a>
+                                <a href="deleteImage.php?id=<?= $PicData['name'] ?>">
+                                    <font color="blue">Удалить</font>
+                                </a>
+                            </td>
+                        <?php }
+                        if ($PicData['user_id'] !== SessionManager::create()->user('user_id')) { ?>
+                            <td>
+                                <a href="lk_form.php?id=<?= $rowUsersData['user_id'] ?>">
+                                    <font color="#f0f8ff"><?= $rowUsersData['user_login'], " " ?><font>
+                                </a>
+                                <a href="deleteImage.php?id=<?= $PicData['name'] ?>">
+                                    <font color="red">Удалить</font>
+                                </a>
+                            </td>
+                        <?php }
+                    }
+                } ?>
                 </thead>
                 <tbody align="center" bgcolor="black">
                 <tr>
@@ -251,7 +264,8 @@ class imagesContorller
                         <?php foreach ($allViewsData as $rowViewData) {
                             if ($rowViewData['name'] == $PicData['name']) { ?>
                                 <p id="views" style="color: white">Просмотров: <?= $rowViewData['views'] ?></p>
-                            <?php } } ?>
+                            <?php }
+                        } ?>
                     </td>
                 </tr>
                 </tbody>
